[영업관리/인사현황] - 엑셀 관련 처리 및 인사 차트 수정 - 숫자관련 Long으로 변경

+++ Loopin/src/main/java/com/itwillbs/c4d2412t3p1/entity/OrderDetail.java
@@ -39,7 +39,7 @@ public class OrderDetail {
     private Long material_am;
 
     @Column(name = "order_ct")
-    private BigDecimal order_ct;
+    private Long order_ct;
 
     @Column(name = "order_ed")
     private String order_ed;
